baseline/gmdr/requirements.txt
http://download.pytorch.org/whl/cu80/torch-0.3.0.post4-cp35-cp35m-linux_x86_64.whl
pyyaml
six
torchtext==0.2.1
future
tensorboardX
progressbar2
sklearn
